+++ packages/adminui-vue/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.6.6]
+### Fixed
+ - limit error dialog width to 250px to prevent text output outside viewport for errors what appears on the right border
+
 ## [1.6.5]
 ### Changed
  - **Breaking** `u-input` v-model support is removed, now gets and sets value from store automatically